Changed some errors in reduction

+++ Reduction_bank_conflicts/sum_cuda.c
@@ -31,7 +31,7 @@ void init(REAL *A, int n)
 {
     int i;
     for (i = 0; i < n; i++) {
-        A[i] = (double)drand48();
+        A[i] = (float)drand48();
     }
 }
